+++ ElvUI_MerathilisUI/Core/Changelog/6.23.lua
@@ -1,8 +1,16 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[623] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
-	NEW = {},
-	IMPROVEMENTS = {},
+	RELEASE_DATE = "28.07.2024",
+	FIXES = {
+		"[Objective Tracker]: Fix an error on the Scenario Tracker",
+	},
+	NEW = {
+		"[Install/Layout]: Enable Rectangle Minimap by default, you might need to adjust your layout a bit",
+	},
+	IMPROVEMENTS = {
+		"[WorldMap]: Updated RawMap Data for 11.0",
+		"[Instance Difficulty]: Updated Logic from Instance Difficutly",
+		"[Skins]: Updated some skin",
+	},
 }