+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Changes ###
 
-  * [Announcement]: Added back some Options, whoooopppsiii
-  * [Objective Tracker]: A lot of work on the Objective Tracker. Thx fang2hou and Blinkii
-  * [Widget Skin]: Some gradient look on the Widget Skin
+  * [Objective Tracker]: Fix an error on the Scenario Tracker
+  * [Install/Layout]: Enable Rectangle Minimap by default, you might need to adjust your layout a bit
+  * [WorldMap]: Updated RawMap Data for 11.0
+  * [Instance Difficulty]: Updated Logic from Instance Difficutly
+  * [Skins]: Updated some skin